+++ zone/lua_bot_impl.cpp
@@ -1,0 +1,80 @@
+#include "bot.h"
+#include "lua_bot.h"
+#include "lua_iteminst.h"
+#include "lua_mob.h"
+
+void Lua_Bot::AddBotItem(uint16 slot_id, uint32 item_id) {
+	Lua_Safe_Call_Void();
+	self->AddBotItem(slot_id, item_id);
+}
+
+void Lua_Bot::AddBotItem(uint16 slot_id, uint32 item_id, int16 charges) {
+	Lua_Safe_Call_Void();
+	self->AddBotItem(slot_id, item_id, charges);
+}
+
+void Lua_Bot::AddBotItem(uint16 slot_id, uint32 item_id, int16 charges, bool attuned) {
+	Lua_Safe_Call_Void();
+	self->AddBotItem(slot_id, item_id, charges, attuned);
+}
+
+void Lua_Bot::AddBotItem(uint16 slot_id, uint32 item_id, int16 charges, bool attuned, uint32 augment_one) {
+	Lua_Safe_Call_Void();
+	self->AddBotItem(slot_id, item_id, charges, attuned, augment_one);
+}
+
+void Lua_Bot::AddBotItem(uint16 slot_id, uint32 item_id, int16 charges, bool attuned, uint32 augment_one, uint32 augment_two) {
+	Lua_Safe_Call_Void();
+	self->AddBotItem(slot_id, item_id, charges, attuned, augment_one, augment_two);
+}
+
+void Lua_Bot::AddBotItem(uint16 slot_id, uint32 item_id, int16 charges, bool attuned, uint32 augment_one, uint32 augment_two, uint32 augment_three) {
+	Lua_Safe_Call_Void();
+	self->AddBotItem(slot_id, item_id, charges, attuned, augment_one, augment_two, augment_three);
+}
+
+void Lua_Bot::AddBotItem(uint16 slot_id, uint32 item_id, int16 charges, bool attuned, uint32 augment_one, uint32 augment_two, uint32 augment_three, uint32 augment_four) {
+	Lua_Safe_Call_Void();
+	self->AddBotItem(slot_id, item_id, charges, attuned, augment_one, augment_two, augment_three, augment_four);
+}
+
+void Lua_Bot::AddBotItem(uint16 slot_id, uint32 item_id, int16 charges, bool attuned, uint32 augment_one, uint32 augment_two, uint32 augment_three, uint32 augment_four, uint32 augment_five) {
+	Lua_Safe_Call_Void();
+	self->AddBotItem(slot_id, item_id, charges, attuned, augment_one, augment_two, augment_three, augment_four, augment_five);
+}
+
+void Lua_Bot::AddBotItem(uint16 slot_id, uint32 item_id, int16 charges, bool attuned, uint32 augment_one, uint32 augment_two, uint32 augment_three, uint32 augment_four, uint32 augment_five, uint32 augment_six) {
+	Lua_Safe_Call_Void();
+	self->AddBotItem(slot_id, item_id, charges, attuned, augment_one, augment_two, augment_three, augment_four, augment_five, augment_six);
+}
+
+uint32 Lua_Bot::CountBotItem(uint32 item_id) {
+	Lua_Safe_Call_Int();
+	return self->CountBotItem(item_id);
+}
+
+Lua_Mob Lua_Bot::GetOwner() {
+	Lua_Safe_Call_Class(Lua_Mob);
+	return Lua_Mob(self->GetOwner());
+}
+
+bool Lua_Bot::HasBotItem(uint32 item_id) {
+	Lua_Safe_Call_Bool();
+	return self->HasBotItem(item_id);
+}
+
+void Lua_Bot::RemoveBotItem(uint32 item_id) {
+	Lua_Safe_Call_Void();
+	self->RemoveBotItem(item_id);
+}
+
+Lua_ItemInst Lua_Bot::GetBotItem(uint16 slot_id) {
+	Lua_Safe_Call_Class(Lua_ItemInst);
+	return self->GetBotItem(slot_id);
+}
+
+uint32 Lua_Bot::GetBotItemIDBySlot(uint16 slot_id) {
+	Lua_Safe_Call_Int();
+	return self->GetBotItemBySlot(slot_id);
+}
+
